nsdispatch removal from gethostbyname()
This gets rid of several layers of obfuscation in between nsdispatch()
and the functions which actually lookup for a hostname in the hosts
file and dns.

This opens the way for even more cleanups, which I will leave for
another day; let's first rip off nsdispatch() from everywhere, which
lets us kill another dependency on an internal bionic header.

Test: atest netd_integration_test

Change-Id: I911fb254939ecc9661695d0ca5474c3caeb4c1d9