Fix hwasan stack buffer overflow in resolver stats
This change will handle netdID unknown case to avoid accessing the uninitialized
local variable stats. If the netId is unknown (which can happen for
valid net IDs for which no DNS servers have yet been configured),
it need to return error directly instead of access the stats.
(The happening is if stats->sample_next is uninitialized or corrupted.)


Test: ./runtests.sh pass
Bug: 123593418

Change-Id: I76b8b23ff618071e3aea9d8abc8aff8887bbab88