Move ExternalPrivateDnsStatus away from libnetd_resolv's public headers
This is a first pass at sanitizing the public headers of libnetd_resolv
by moving out things that are only used internally.

The main motivation is reducing noise in ABI dumps generated for the ABI
checker. It's also a nice opportunity to cut off the C wrappers that we
no longer need.

Test: cd system/netd && atest
Change-Id: I08689c263c4fd96363666a108ba08d768ad327df