Rename frozen netd AIDL API directories.
Currently, all the netd AIDL interfaces specify an api_dir of
"binder/api". This does not work because the compiler does not
support using the same api_dir for multiple aidl_interface rules.

This affects only netd_event_listener_interface and
dnsresolver_aidl_interface. netd_event_listener_interface is
unaffected because it is not yet frozen.

Test: m
Bug: 133124190
Change-Id: I42ea79194d0d3a1126fd513ed39d73fc236aaca3