dns_responder: Add an optional DNS header mapping to build a response
Currently, the DNS response is built from the registered hostname or
address mapping. The response header fields have been predefined mostly
and that may limit the test coverage indirectly.

An alternative way is to add an optional DNS header mapping to build
a response. The response could be built from a preconfigured DNS
header if necessary. That helps to extend the test coverage.

Test: cd packages/modules/DnsResolver && atest
Change-Id: Ib56ae7b076e8e1e9ebbbca187b66ee91643dd0b0