Add private DNS validation thread tracker
The behavior is that the different dns changed in 10ms every 3-5s,
and this dns vibrate will be set by DnsResolver. In ResolverController,
will start one thread for validate each dns server. If dns server removed,
and the trace of dns server will be erased too, but dns validate thread
not destory immediately.
So for this scene will start new thread for dns vibrate in 10ms,
and many thread will cause OOM.

Test:
We verified this solution in that special user environment and merged it
at end of Feb 2019.
We didn't get any regression issue report so far (Sep 2019)

Bug: 139782337

Change-Id: I88431d77f575ccffa5af7c596a915f223e515027