Add stats samples by looking up server address
When adding a stats sample, it passes the server index rather than
the server address. It's not wrong if the order of dns servers in
lookup threads is fixed. However, it restrains the flexibility to
change the order as well.

This change fixes it to pass the server address so that the order
of dns servers in a lookup thread can change.

Bug: 137169582
Test: atest --include-subdirs packages/modules/DnsResolver
Change-Id: I1f4542165e90289efe8bf6d618639b82a8cc740e