Delete the _u._ext nonsense from res_state
Also eliminate _u._ext.nscount, which was nothing but a flag to
clear _u._ext.nssocks on the first run of res_nsend(). Moved the
initialization to res_init() instead.

We're starting to get closer to the final shape of this code. Next steps
would involve eliminating res_static and all the code in res_state.cpp
dealing with thread-local storage.

Test: m, flash, atest
Change-Id: I4585b9d798d5ff9b8482dc12619d998d77398a0b