Change DnsTlsTransport to store a copy of a query
This is preparation for configurable per-query timeout in order to
allow a lookup thread to exit while a DnsTlsSocket loop thread is
still waiting the response. Since a DnsTlsTransport instance is
accessible to both lookup threads and a DnsTlsSocket loop thread, change
it to store a copy of a query in case of any operation needed for
the query in the loop thread; for example, the response arrives after
timeout.

Bug: 120182528
Bug: 141218721
Test: cd packages/modules/DnsResolver && atest
Change-Id: I57bf40cb1eb0d688d5328242649c841f20d81430