Support multinetwork tests
The integration test used to set up testing DNS servers on loopback
interface. To support testing functionality for multinetwork, make
the test able to send queries to a TUN interface, and the queries
will be forwarded to the testing DNS servers.

To forward packets, implement a forwarder which can translate packets
(v4-to-v4 or v6-to-v6) between the resolver and testing DNS servers
and can forward packets to each other.

Also add three tests:
  GetAddrInfo_AI_ADDRCONFIG
  NetworkDestroyedDuringQueryInFlight
  OneCachePerNetwork

And remove unused libraries from the test:
  libnetd_test_tun_interface
  libnetd_test_utils

Test: cd packages/modules/DnsResolver && atest
Change-Id: I52a52ce59373bc8b9462064c0409b657696c379f