Fix flaky test RepeatedSetup_NoRedundantPrivateDnsValidation
There is a time gap between the time when PrivateDnsConfiguration
reports onPrivateDnsValidationEvent and the time when
PrivateDnsConfiguration updates its map which is used for tracking
validation threads. If the test starts a subsequent setup (which expects
the resolver to begin a new validation) before PrivateDnsConfiguration
updates the map, the expected validation won't start because
PrivateDnsConfiguration thinks the previous validation hasn't finished.

This change is a workaround which adds waiting time to avoid the timing
issue, and the test running time increases 300ms.

Bug: 150678049
Bug: 151990391
Test: resolv_integration_test passed
Change-Id: I176d7d20b7edc79b37b954602961c149aef64348