dnsresolver_aidl_interface is marked as available to platform
The AIDL interface is depended on by services.net which is part of the
platform. With b/153073816, we will be checking the availability to the
platform; i.e. platform-available module dependinng on
platform-unavailable module will cause a build error.

Prepare for the change by correctly marking that the interface is also
available to the platform.

Bug: 153073816
Test: m

Change-Id: I69e8ebf06e71918f430c70ae1411d3bcded9f464