Test for unexpected prefix updates in SetAndClearNat64Prefix.
Add code to check that no onNat64PrefixEvent calls were sent by
the resolver other than the ones expected by the test.

This adds a simple counter of unexpected events to the test
DnsMetricsListener. It does not offer the ability to check which
events occur and in what order, but it does at least not require
adding timeouts to ensure that the unexpected events do not
occur.

Bug: 156914456
Test: test-only change
Change-Id: Ic86f93161293b4ad5366fd9cab2e26d4a6f636c4