Make DoT retries configurable
DnsTlsTransport re-issues pending queries when onClosed is called.
The call to onClosed is triggered when 1) asynchronous handshake
fails and 2) SSL socket idles for 20 seconds. In either case, retry
on the same DoT server is not always a good solution. Instead, there
are some considerable options, like trying next DoT server, fallbacking
to Do53, or simply returning query failure.

Tuning DoT retries is especially significant to asynchronous
handshake feature because the timeout of the feature is calculated as:

  timeout = dot_connect_timeout_ms * dot_maxtries

Bug: 149445907
Test: cd packages/modules/DnsResolver
      atest with combination of (dot_async_handshake, dot_maxtries)
      which are (0, 3), (0, 1), (1, 3), and (1, 1)
Change-Id: Iceb7bc7f0f6736d900384d1a11eea470761ee32c