Be explicit about binderCallLogFn
As binderCallLogFn is changed to be template function, std::bind() and
placeholder::_1 don't work well about type deduction.
To make C++ compiler happy, switched to lambda.

Note that the binderCallLogFn change is about the AIDL's gen_log feature
change that callback is given a struct rather than a JSON object.

Bug: 172188290
Test: m
Test: atest resolv_integration_test
Change-Id: I0097738713416eea2b5d47b73fadd8cc2c8ee5e4