Do not install tethering in TEST_MAPPING
The current tethering module has in-progress changes in non-mainline
branches, so it cannot be installed.
Disable the tests in non-mainline branches, considering that they are
still run in mainline branches, which will contain the same tethering
code, but without the connectivity artifacts added to the tethering
apex.

Bug: 177290955
Test: TEST_MAPPING needs to be tested on the infra
Ignore-AOSP-First: This change must not go into AOSP
Change-Id: I83243474f8e2f72d03ef68176053f9db793d127b