DO NOT MERGE: Add a soong var that will be used to enable/disable rust relevant targets
Add a soong variable which will be used to enable/disable rust
relevant targets. This makes it easy to enable or disable
the targets all together. It also makes it much easier to
automerge Android.bp changes from AOSP to mainline-prod without
merge conflict, because the build target definitions are always
the same, only the variable at the top of the file changes.

Bug: 155855709
Test: m
Ignore-AOSP-First: needed to solve merge conflicts and rust toolchain
problem.

Change-Id: I21945b25e3e4e04e6110030d45b82fa1acf6c409