Update the default sampling flag of DNS resolver metrics events
The sample rate of DNS resolver metrics events depends on Experiment
flag and default sampling flag. Currently, the sampling experiment
has been completed. Therefore update the new default sampling flag to
reduce the metrics event size.

1. Current default sampling rate:
   DEFAULT_SUBSAMPLING_MAP[] = "default:1 0:100 7:10
   metrics data size: 170k byte per device
2. Old Experiment:
   DEFAULT_SUBSAMPLING_MAP[] = "default:5 0:300 2:80 7:80"
   metrics data size: 11k byte per device
3. New default sampling rate:
   DEFAULT_SUBSAMPLING_MAP[] = "default:8 0:400 2:110 7:110"
   metrics data size: 7.9k byte per device
4. New Experiment:(including the temporary fix of b/185097438)
   DEFAULT_SUBSAMPLING_MAP[] = "default:8 0:400 2:110 4:110 7:110"
   metrics data size: 7.9k byte per device

Because there is already a solution for b/185097438 in the code.
Therefore, what this commit will do is item3 to reduce the size
of the event.

Bug: 184830442
Test: atest resolv_unit_test
Change-Id: I3c0ab4a4504dfd5f8c789348587de33e01951920