Revert "Read test cacert and keys from files"
This reverts commit af130a4c29ed5adcb83463e9ac2e61429cb4c84c.

When we build resolv_{unit,integration,stres,gold}_test, testing
pem files are copied and pushed to corresponding output directory.

The following build command moves all of the pem files to out/dist
and thinks that the files are duplicated.

m TARGET_BUILD_APPS="resolv_unit_test resolv_integration_test" dist
(This is a simplified command. The actual command which caused
the issue is in b/193063575)

We will use pipe as a workaround to read the testing certificates.

Bug: 193063575
Test: m TARGET_BUILD_APPS="com.google.android.resolv resolv_unit_test
      resolv_integration_test resolv_gold_test resolv_stress_test" dist
Change-Id: I0553aa4874603089630e771298b16c96946a3c2a