DoH: Increase the size of Network mpsc channel
Increase the channel size so that the DoH client can queue more
DNS requests before a DoH connection is established.

Apart from that, since the size of the queue is increased, it can
take more time to drain all the DNS requests out of the queue,
especially when DoH connections can't be established. This change
also add the check for expired DNS requests in Network object
before send them to Connection object.

Bug: 207301204
Test: cd packages/modules/DnsResolver && atest
Change-Id: Icfcbc7d389621cbbac23b5880907c26f1ab2f57f