Test: Use 8-bytes of dcid as connection ID
After this change, test DoH server will use the first 8 bytes of
ocid as the new connection ID. This change is a preparation for
0-RTT in order to identify 0-RTT packets.

The handshake flow is illustrated below. We used to generate a
random S3. This works in 1-RTT handshake since the client changes
to use S3 as ocid starting from the second packet. However, in
a 0-RTT handshake, the client sends 0-RTT packets followed by the
Initial packet with the same dcid. In order to recognize the
0-RTT packets that belong to the client, we use the first 8-bytes
of S1 as S3. Hence, we know that packets (DCID=S1) and packets
(DCID=S3) belong to the same QUIC connection.

1-RTT Handshake:
  Client                                             Server
  Initial: DCID=S1, SCID=C1 ->
                               <- Initial: DCID=C1, SCID=S3
                             ...
  SHORT: DCID=S3

0-RTT Handshake:
  Client                                             Server
  Initial: DCID=S1, SCID=C1
  0-RTT: DCID=S1, SCID=C1   ->
                               <- Initial: DCID=C1, SCID=S3
			     ...
  SHORT: DCID=S3

Bug: 235763732
Test: atest
Change-Id: I7178b9244c8582c8282fc9ce15b7492511c1f8b1