Setting apex version to the placeholder value
When the apex version is set to 0, the version number will be replaced by the build system to the default value for the current branch.
The default value for mainline module manifest version is stored in the defaultManifestVersion constant: https://android.googlesource.com/platform/build/soong/+/refs/heads/master/apex/constants.go#36

The version should not be changed manually in the manifest. Release tooling will set this version to the release version on the release branches.

Test: presubmit
Bug: 231691643
Change-Id: I3c89d02f9d6eafa34aa3bafc66063e677faddb16
Merged-In: I30b5b486bca521ebdeed17dba7afc61f2c529ab8