Test: Add connectivity check for DoHFFITest SmokeTest
The test will be skipped if there is no network connected and it
will connect to IPv4 or IPv6 server based on the connectivity.

Probe timeout and idle timeout are extended to 10 seconds.

So far, doh_ffi_test isn't in presubmit, so we need to specify
the test name (atest doh_ffi_test) to run it.

Bug: 237767883
Test: atest doh_ffi_test
Change-Id: I7d2977acc5c80cb960ce510a4d3bb53c9628d033