Set IPv6 route for mdns tests
Just as we added a v4 route for mdns tests, a v6 routes is also
required. It can pass the test previously is because of the existence
of mdnsd. If the mdnsd is terminated by other tests, such as
CtsNetTestCasesMaxTargetSdk30, and netd_integration_test is run after
them, the .local tests in the DNS resolver will fail.

Bug: 240888834
Test: atest CtsNetTestCasesMaxTargetSdk30 resolv_integration_test
Change-Id: Ib7e91a19077b5ff1d87bd429c82440cfaf870470